requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.2.0
Django==5.0.1
dj-database-url==2.3.0
djangorestframework==3.16.0
django-filter==24.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.2
uvicorn==0.34.2
whitenoise==6.9.0
